// eshop-main/src/graphql/__generated__/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
// ...
export type Order = {
  __typename?: 'Order';
  id: string;
  orderNumber: string;
  status: string;
  total: number;
  createdAt: string;
};

export type OrderConnection = {
  __typename?: 'OrderConnection';
  edges: Array<{ __typename?: 'OrderEdge'; node: Order }>;
};
// ...
export type GetOrdersQuery = { orders?: Maybe<OrderConnection> };

export type GetOrdersQueryVariables = Exact<{ [key: string]: never }>;
// ...
export const GetOrdersDocument = gql`
  query GetOrders {
    orders {
      edges {
        node {
          id
          orderNumber
          status
          total
          createdAt
        }
      }
    }
  }
` as unknown as TypedDocumentNode<GetOrdersQuery, GetOrdersQueryVariables>;